requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
requests==2.31.0
openpyxl==3.1.2
APScheduler==3.10.4
python-dotenv==1.0.0